+++ 2019/sol/include/sol.h
@@ -8,8 +8,8 @@ extern "C"
 unsigned long solve_2019_1_1(char const *input_filename);
 long solve_2019_1_2(char const *input_filename);
 
-int solve_2019_2_1(char const *input_filename);
-int solve_2019_2_2(char const *input_filename);
+long solve_2019_2_1(char const *input_filename);
+long solve_2019_2_2(char const *input_filename);
 
 int solve_2019_3_1(char const *input_filename);
 int solve_2019_3_2(char const *input_filename);
@@ -17,13 +17,13 @@ int solve_2019_3_2(char const *input_filename);
 int solve_2019_4_1(char const *input_filename);
 int solve_2019_4_2(char const *input_filename);
 
-int solve_2019_5(char const *input_filename, int input);
+long solve_2019_5(char const *input_filename, int input);
 
 int solve_2019_6_1(char const *input_filename);
 int solve_2019_6_2(char const *input_filename);
 
-int solve_2019_7_1(char const *input_filename);
-int solve_2019_7_2(char const *input_filename);
+long solve_2019_7_1(char const *input_filename);
+long solve_2019_7_2(char const *input_filename);
 
 int solve_2019_8_1(char const *input_filename, int width, int height);
 char *solve_2019_8_2(char const *input_filename, int width, int height);
